Cleanup classes and tests in the io.cassandrareaper.jmx pacakge
 - introduce the JmxProxy capability, hiding the rest of the implementation
 - correct the test package name,
 - hide JmxConnectionsInitializer's constructor
 - hide StorageServiceMBean20
