+++ src/server/src/main/java/io/cassandrareaper/ReaperApplication.java
@@ -294,10 +294,8 @@ public final class ReaperApplication extends Application<ReaperApplicationConfig
 
   private void initializeJmxSeedsForAllClusters() {
     LOG.info("Initializing JMX seed list for all clusters...");
-    try (JmxConnectionsInitializer jmxConnectionsIntializer =
-            new JmxConnectionsInitializer(context);
-        Timer.Context cxt =
-            context
+    try (JmxConnectionsInitializer jmxConnectionsIntializer = JmxConnectionsInitializer.create(context);
+        Timer.Context cxt = context
                 .metricRegistry
                 .timer(MetricRegistry.name(JmxConnectionFactory.class, "jmxConnectionsIntializer"))
                 .time()) {
